Parallel_loop_reduction_add_general_type_check_pt1.cpp:
	From the OpenACC 3.3 spec (2.5.15, “reduction clause”):
Line 1231: a var in a reduction clause must be a scalar variable name, an aggregate variable name, an array element, or subarray (refer to section 2.7.1). Line 1225-1227: However, for each reduction operator, the supported data types include only the types permitted as operands to the corresponding operator in the base language. So while it states bools are supported in line 1223, weather bool works for a specific operator like + depends on weather + is valid on bool in the language. But using + with bool is implementation dependent and not required to be supported because of type promotion bool + bool -> int. This is why, even though the spec includes bool, compilers like NVIDIA's nvc might reject or mishandle reductions over bool with +.

The original test1() used bool types in an OpenACC reduction. According to the OpenACC 3.3 specification (2.5.15), reductions are only required to be supported for arithmetic and compatible user-defined types. Since bool is a logical type, using it in a reduction results in implementation-dependent or undefined behavior. Furthermore, the functionality of test1() overlaps entirely with test6(), which tests reductions with int. To avoid redundancy and non-conforming behavior, test1() has been removed.

	-After change to int: PASSED. <- (this was just for testing bc test 6 tests for ints)
	-Also after bool is completely removed it PASSED