Correct set_if tests
This patch fixes two problems with the set_if tests:

First, as acc_get_device_type is a function, the comparison was a
function-pointer to integer comparison inside the 'if'.  This operation
is ill-formed in C++.  However, this patch changes for both the C and
C++ test to make sure we call the function instead.

Second, OpenACC's restrictions in the 'set' section (2.14.3 in
    OpenACC3.3) has a restriction of:
"At least one default_async, device_num, or device_type clause must
appear".
This patch adds default_async with acc_async_default to both uses of
'set' in both tests.