Correct use of 'restrict' in C++ tests.
'restrict' is only a keyword in C, and it is not a conforming extension
to support it in C++, so GCC and Clang both diagnose this. However, most
compilers expose '__restrict' as an alias for this in a conforming
manner.

This patch changes the uses of 'restrict' in C++ tests to '__restrict'
so that these tests can compile.