Fix two issues in parallel_create_zero
These tests have a few issues, but this patch fixes two of them:

First, it adds the close paren on the 'copyout' clause that was
missing.

Second, it corrects the capitalization on the call to 'Test1', which was
resulting in the test not being run properly.

This does NOT fix the ill-formed use of `loop` however.